@@ -24,9 +24,11 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 """
 
-from typing import Any, Callable
+from typing import Any, Callable, Dict
 
-from pytato.array import Array
+from pytato.array import (
+        Array, IndexLambda, Namespace, Placeholder, MatrixProduct, Stack,
+        Roll, AxisPermutation, Slice, DataWrapper, SizeParam, DictOfNamedArrays)
 
 __doc__ = """
 .. currentmodule:: pytato.transform
@@ -34,7 +36,9 @@ __doc__ = """
 Transforming Computations
 -------------------------
 
-.. autoclass:: Mapper
+.. autoclass:: CopyMapper
+.. autofunction:: copy_namespace
+.. autofunction:: copy_dict_of_named_arrays
 
 """
 
@@ -59,11 +63,11 @@ class Mapper:
         raise ValueError("%s encountered invalid foreign object: %s"
                 % (type(self).__name__, repr(expr)))
 
-    def __call__(self, expr: Array, *args: Any, **kwargs: Any) -> Any:
+    def rec(self, expr: Array, *args: Any, **kwargs: Any) -> Any:
         method: Callable[..., Array]
 
         try:
-            method = getattr(self, expr.mapper_method)
+            method = getattr(self, expr._mapper_method)
         except AttributeError:
             if isinstance(expr, Array):
                 return self.handle_unsupported_array(expr, *args, **kwargs)
@@ -72,9 +76,116 @@ class Mapper:
 
         return method(expr, *args, **kwargs)
 
-    rec = __call__
+    def __call__(self, expr: Array, *args: Any, **kwargs: Any) -> Any:
+        return self.rec(expr, *args, **kwargs)
+
+
+class CopyMapper(Mapper):
+
+    def __init__(self, namespace: Namespace):
+        self.namespace = namespace
+        self.cache: Dict[Array, Array] = {}
+
+    def rec(self, expr: Array) -> Array:  # type: ignore
+        if expr in self.cache:
+            return self.cache[expr]
+        result: Array = super().rec(expr)
+        self.cache[expr] = result
+        return result
+
+    def map_index_lambda(self, expr: IndexLambda) -> Array:
+        bindings = {
+                name: self.rec(subexpr)
+                for name, subexpr in expr.bindings.items()}
+        return IndexLambda(namespace=self.namespace,
+                expr=expr.expr,
+                shape=expr.shape,
+                dtype=expr.dtype,
+                bindings=bindings,
+                tags=expr.tags)
+
+    def map_placeholder(self, expr: Placeholder) -> Array:
+        return Placeholder(namespace=self.namespace,
+                name=expr.name,
+                shape=expr.shape,
+                dtype=expr.dtype,
+                tags=expr.tags)
+
+    def map_matrix_product(self, expr: MatrixProduct) -> Array:
+        return MatrixProduct(x1=self.rec(expr.x1),
+                x2=self.rec(expr.x2),
+                tags=expr.tags)
+
+    def map_stack(self, expr: Stack) -> Array:
+        arrays = tuple(self.rec(arr) for arr in expr.arrays)
+        return Stack(arrays=arrays, axis=expr.axis, tags=expr.tags)
+
+    def map_roll(self, expr: Roll) -> Array:
+        return Roll(array=self.rec(expr.array),
+                shift=expr.shift,
+                axis=expr.axis,
+                tags=expr.tags)
+
+    def map_axis_permutation(self, expr: AxisPermutation) -> Array:
+        return AxisPermutation(array=self.rec(expr.array),
+                axes=expr.axes,
+                tags=expr.tags)
+
+    def map_slice(self, expr: Slice) -> Array:
+        return Slice(array=self.rec(expr.array),
+                begin=expr.begin,
+                size=expr.size,
+                tags=expr.tags)
+
+    def map_data_wrapper(self, expr: DataWrapper) -> Array:
+        return DataWrapper(namespace=self.namespace,
+                name=expr.name,
+                data=expr.data,
+                shape=expr.shape,
+                tags=expr.tags)
+
+    def map_size_param(self, expr: SizeParam) -> Array:
+        return SizeParam(self.namespace, name=expr.name, tags=expr.tags)
 
 # }}}
 
+
+# {{{ mapper frontends
+
+def copy_namespace(source_namespace: Namespace,
+        copy_mapper: CopyMapper) -> Namespace:
+    """Copy the elements of *namespace* into a new namespace.
+
+    :param source_namespace: The namespace to copy
+    :param copy_mapper: A mapper that performs copies into a new namespace
+    :returns: A new namespace containing copies of the items in *source_namespace*
+    """
+    for name, val in source_namespace.items():
+        mapped_val = copy_mapper(val)
+        if name not in copy_mapper.namespace:
+            copy_mapper.namespace.assign(name, mapped_val)
+    return copy_mapper.namespace
+
+
+def copy_dict_of_named_arrays(source_dict: DictOfNamedArrays,
+        copy_mapper: CopyMapper) -> DictOfNamedArrays:
+    """Copy the elements of a :class:`~pytato.DictOfNamedArrays` into a
+    :class:`~pytato.DictOfNamedArrays` with a new namespace.
+
+    :param source_dict: The :class:`~pytato.DictOfNamedArrays` to copy
+    :param copy_mapper: A mapper that performs copies into a new namespace
+    :returns: A new :class:`~pytato.DictOfNamedArrays` containing copies of the
+        items in *source_dict*
+    """
+    if not source_dict:
+        return DictOfNamedArrays({})
+
+    data = {}
+    copy_namespace(source_dict.namespace, copy_mapper)
+    for name, val in source_dict.items():
+        data[name] = copy_mapper(val)
+    return DictOfNamedArrays(data)
+
+# }}}
 
 # vim: foldmethod=marker
