@@ -1,8 +1,6 @@
-__copyright__ = """
-Copyright (C) 2020 Andreas Kloeckner
-Copyright (C) 2020 Matt Wala
-Copyright (C) 2020 Xiaoyu Wei
-"""
+#!/usr/bin/env python
+
+__copyright__ = "Copyright (C) 2020 Andreas Kloeckner"
 
 __license__ = """
 Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -24,29 +22,36 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 """
 
-from pytato.array import (
-        Namespace, Array, DictOfNamedArrays, Tag, UniqueTag,
-        DottedName, Placeholder, IndexLambda,
+import sys
 
-        make_dict_of_named_arrays,
-        make_placeholder, make_size_param, make_data_wrapper,
+import numpy as np
+import pyopencl as cl  # noqa
+import pyopencl.array as cl_array  # noqa
+import pyopencl.cltypes as cltypes  # noqa
+import pyopencl.tools as cl_tools  # noqa
+from pyopencl.tools import (  # noqa
+        pytest_generate_tests_for_pyopencl as pytest_generate_tests)
+import pytest  # noqa
 
-        matmul, roll, transpose, stack,
-        )
+import pytato as pt
 
-from pytato.codegen import generate_loopy
-from pytato.target import Target, PyOpenCLTarget
 
-__all__ = (
-        "DottedName", "Namespace", "Array", "DictOfNamedArrays",
-        "Tag", "UniqueTag", "Placeholder", "IndexLambda",
+def test_namespace_copy():
+    namespace = pt.Namespace()
+    x = pt.Placeholder(namespace, "x", (5,), np.int)
+    namespace.assign("xsquared", x * x)
 
-        "make_dict_of_named_arrays", "make_placeholder", "make_size_param",
-        "make_data_wrapper",
+    namespace_copy = namespace.copy()
+    assert len(namespace_copy) == 2
+    assert isinstance(namespace_copy["x"], pt.Placeholder)
+    assert isinstance(namespace_copy["xsquared"], pt.IndexLambda)
 
-        "matmul", "roll", "transpose", "stack",
 
-        "generate_loopy",
+if __name__ == "__main__":
+    if len(sys.argv) > 1:
+        exec(sys.argv[1])
+    else:
+        from pytest import main
+        main([__file__])
 
-        "Target", "PyOpenCLTarget",
-)
+# vim: filetype=pyopencl:fdm=marker
